Refactor the filesystem (#63)
* Rename sans_fitter.py to fitter.py

Removes the package-name echo (sans_fitter.sans_fitter.SANSFitter).
Public API is unchanged; SANSFitter and get_all_models are still
re-exported from the package root.


* Move data loading and arithmetic into a data/ subpackage

data_loader.py becomes data/loader.py and data_ops.py becomes
data/ops.py. _has_real_data was imported by five sibling modules, so it
was never really private; it is now public as has_real_data.

sans_fitter.data_ops remains importable as an alias of sans_fitter.data.ops,
so 'from sans_fitter import data_ops' keeps working in examples and
notebooks.


* Group model configuration into a modeling/ subpackage

parameter_manager.py, polydispersity.py and structure_factor.py form a
closed cluster (only ParameterManager ties them together), so they move
to modeling/parameters.py, modeling/polydispersity.py and
modeling/structure_factor.py. Top-level re-exports are unchanged.


* Split pr_inversion.py into an inversion/ subpackage

The 1366-line module had four separable layers, now four modules:
basis.py (sine-basis functions and transforms), solver.py (data
preparation, matrix assembly, SVD solve, invert()), estimate.py
(alpha/n_terms heuristics, auto_invert, D_max scans) and result.py
(PrResult and the error types).

sans_fitter.pr_inversion remains importable as an alias of
sans_fitter.inversion, so existing 'from sans_fitter import
pr_inversion' usage keeps working. Tests that reached module-private
helpers now import the specific submodule.


* Merge contracts.py into results.py

contracts.py held a single 27-line dataclass (ParameterStateSnapshot)
serving the same role as results.py: shared types passed between
layers. One home for both removes a file without losing clarity.


* Address refactor review: doc sweep, test fixes, API-stability test

- Update prose and docstring module references to the canonical paths
  (sans_fitter.data.ops, sans_fitter.inversion) in docs/usage.md, the
  example scripts, examples.py, fitter.py and the two demo notebooks.
  Import statements keep using the stable top-level aliases.
- Fix the test_has_real_data_* method names that a mechanical rename had
  mangled to testhas_real_data_*.
- Capture and patch _invert_prepared on the same module
  (inversion.estimate) in the explore_dmax failure tests.
- Comment the frozen 'sans_fitter.data_ops' provenance string so it is
  not 'fixed' into a compatibility break later.
- Add tests/test_public_api.py locking in __all__ importability and the
  data_ops/pr_inversion module aliases.


* another round of code reviews

---------
