+++ src/sans_fitter/__init__.py
@@ -2,12 +2,14 @@
 SANS Model Fitter - A flexible template for fitting SANS data with SasModels
 """
 
-from . import data_ops, examples, pr_inversion
-from .parameter_manager import ParameterManager
-from .polydispersity import PD_DEFAULTS, PD_DISTRIBUTION_TYPES
-from .pr_inversion import InsufficientDataError, PrEstimationError, PrResult
+from . import examples
+from . import inversion as pr_inversion
+from .data import ops as data_ops
+from .fitter import SANSFitter, get_all_models
+from .inversion import InsufficientDataError, PrEstimationError, PrResult
+from .modeling.parameters import ParameterManager
+from .modeling.polydispersity import PD_DEFAULTS, PD_DISTRIBUTION_TYPES
 from .results import FitResultContract, PosteriorSummary
-from .sans_fitter import SANSFitter, get_all_models
 
 __version__ = '0.3.0'
 __all__ = [
