+++ src/sans_fitter/data/__init__.py
@@ -1,0 +1,12 @@
+"""Data loading, normalization and dataset arithmetic."""
+
+from . import ops
+from .loader import get_fit_index, has_real_data, load_sans_data, normalize_sans_data
+
+__all__ = [
+    'get_fit_index',
+    'has_real_data',
+    'load_sans_data',
+    'normalize_sans_data',
+    'ops',
+]
