+++ src/sans_fitter/modeling/__init__.py
@@ -1,0 +1,14 @@
+"""Model configuration: parameters, polydispersity and structure factors."""
+
+from .parameters import ParameterManager
+from .polydispersity import PD_DEFAULTS, PD_DISTRIBUTION_TYPES, PolydispersityManager
+from .structure_factor import StructureFactorManager, default_parameter_bounds
+
+__all__ = [
+    'PD_DEFAULTS',
+    'PD_DISTRIBUTION_TYPES',
+    'ParameterManager',
+    'PolydispersityManager',
+    'StructureFactorManager',
+    'default_parameter_bounds',
+]
