+++ src/sans_fitter/modeling/structure_factor.py
