+++ src/sans_fitter/results.py
@@ -3,10 +3,35 @@ from typing import Any
 
 import numpy as np
 
-from .data_loader import _has_real_data
+from .data.loader import has_real_data
 
 MIN_POSTERIOR_PARAMETER_COUNT = 2
 MIN_POSTERIOR_SAMPLE_COUNT = 2
+
+
+@dataclass(slots=True)
+class ParameterStateSnapshot:
+    """Read-only parameter state passed into fitting engines.
+
+    All parameter names carried by the snapshot are **canonical sasmodels
+    names** (e.g. ``A_sld`` under a composite model). Any user-facing alias
+    layer is translated away before the snapshot is built.
+    """
+
+    params: dict[str, dict[str, Any]]
+    polydisperse_param_names: list[str]
+    polydisperse_params: dict[str, dict[str, Any]]
+    pd_enabled: bool
+    radius_effective_mode: str
+    structure_factor_name: str | None
+    varying_params: list[str]
+    varying_pd_params: list[str]
+    # Equality links (follower -> target), canonical names. Populated by
+    # link_params() and by the shared= sugar of set_models().
+    linked_params: dict[str, str] = field(default_factory=dict)
+    # Composite-model components as (prefix, moniker, part_model_name)
+    # triples; empty for atomic models.
+    components: tuple[tuple[str, str, str], ...] = ()
 
 
 def _validate_export_lengths(**arrays: Any) -> None:
@@ -180,7 +205,7 @@ class FitResultContract:
         are exported, so every row carries a fitted intensity and residual.
         """
         fitted_curve = self.require_fitted_curve()
-        has_dx = _has_real_data(data.dx)
+        has_dx = has_real_data(data.dx)
 
         index = resolve_fit_index(self.artifacts.fit_index, len(data.x))
         x = np.asarray(data.x)[index]
